Simplify, reduce code repetition, and remove functions
Add Enum for quality choice so 0 and 1 are not spread throughout
without meaning, and reduce number of if..else statements in
relation to it.

Remove unused functions and move related functions to staticmethods.

In class Video, make separate sd_link and hd_link attributes into
dictionary for ease of access with new enum, and add simple
url function to return value from dictionary.