App class enhancements and utils fix
Adjust __init__ to automatically create config.yaml when not found
but only run first time setup when driver_path not found, which
allows for user to have put in their credentials but not
have run the program yet.

Added password field to config.yaml.

Fix in utils.py for checking if link extracted from chromedriver
webpage returns a list.